split app.rs into multiple different pages
