+++ src/app/page/show_page.rs
@@ -1,0 +1,177 @@
+use crate::app::message::Message;
+use crate::app::page::{Page, Pages};
+use crate::app::state::State;
+use crate::app::Lokbuch;
+use crate::frontend;
+use crate::frontend::widgets::header;
+use async_std::task;
+use iced::widget::{button, center, horizontal_space, row, text, text_input, vertical_space};
+use iced::{Center, Element, Fill, Task};
+
+pub struct ShowPage;
+
+impl Page for ShowPage {
+    fn update(&self, lokbuch: &mut Lokbuch, message: Message) -> Task<Message> {
+        match message {
+            Message::Add => {
+                lokbuch.change_page_to(Pages::Add);
+                return text_input::focus("new-lok-name");
+            }
+
+            Message::SearchInputChanged(search_input) => {
+                lokbuch.state.search_input = search_input;
+            }
+
+            Message::Search => {
+                println!("Search pressed!");
+                //implement search by creating a search string made of address, LMname and name
+            }
+
+            Message::ShowLok(id) => {
+                lokbuch.state = State {
+                    selected_lok_id: Some(id),
+                    ..State::default()
+                };
+
+                lokbuch.change_page_to(Pages::Show);
+            }
+
+            Message::Cancel => {
+                lokbuch.change_page_to(Pages::Home);
+            }
+
+            Message::Edit(id) => {
+                let lok = task::block_on(lokbuch.lok_resource_manager.get_lok(id)).expect("Lok does not exist!"); // TODO async edit
+
+                let name_input = lok.name.clone();
+
+                let address_input = if let Some(address) = lok.address.clone() {
+                    address.to_string()
+                } else { String::new() };
+                let lok_maus_name_input = if let Some(lokmaus_name) = lok.lokmaus_name.clone() {
+                    lokmaus_name
+                } else { String::new() };
+                let producer_input = if let Some(producer) = lok.producer.clone() {
+                    producer
+                } else { String::new() };
+                let management_input = if let Some(management) = lok.management.clone() {
+                    management
+                } else { String::new() };
+
+                lokbuch.state = State {
+                    selected_lok_id: Some(id),
+                    name_input,
+                    address_input,
+                    lok_maus_name_input,
+                    producer_input,
+                    management_input,
+                    ..State::default()
+                };
+
+                lokbuch.change_page_to(Pages::Edit);
+                return text_input::focus("new-lok-name");
+            }
+
+            Message::Remove(id) => {
+                task::block_on(lokbuch.lok_resource_manager.remove_lok(id)); // TODO async remove
+
+                lokbuch.change_page_to(Pages::Home);
+            }
+
+            _ => {}
+        }
+        Task::none()
+    }
+
+    fn view<'a>(&self, lokbuch: &'a Lokbuch) -> Element<'a, Message> {
+        let mut lrm = lokbuch.lok_resource_manager.clone();
+        let lok = task::block_on(lrm.get_lok(lokbuch.state.selected_lok_id.unwrap())).expect("lok not found"); // TODO async show
+
+        let header = header(format!("{}", lok.name));
+
+        let left_column = iced::widget::column!(
+                    text("Adresse")
+                    .size(frontend::HEADING_TEXT_SIZE)
+                    .font(frontend::font::bold_font()),
+
+                    text!("{}", lok.get_address_pretty())
+                    .size(frontend::HEADING_TEXT_SIZE),
+
+                    vertical_space(),
+
+                    text("Bezeichnung")
+                    .size(frontend::HEADING_TEXT_SIZE)
+                    .font(frontend::font::bold_font()),
+
+                    text!("{}", lok.name)
+                    .size(frontend::HEADING_TEXT_SIZE),
+
+                    vertical_space(),
+
+                    text("Hersteller")
+                    .size(frontend::HEADING_TEXT_SIZE)
+                    .font(frontend::font::bold_font()),
+
+                    text!("{}", lok.get_producer_pretty())
+                    .size(frontend::HEADING_TEXT_SIZE),
+                ).spacing(10);
+
+        let right_column = iced::widget::column![
+                    text("LOKmaus-Anzeigename")
+                    .size(frontend::HEADING_TEXT_SIZE)
+                    .font(frontend::font::bold_font()),
+
+                    text!("{}", lok.get_lokmaus_name_pretty())
+                    .size(frontend::HEADING_TEXT_SIZE),
+
+                    vertical_space(),
+                    vertical_space(),
+                    vertical_space(),
+                    vertical_space(),
+
+                    text("Bahnverwaltung")
+                    .size(frontend::HEADING_TEXT_SIZE)
+                    .font(frontend::font::bold_font()),
+
+                    text!("{}", lok.get_management_pretty())
+                    .size(frontend::HEADING_TEXT_SIZE)
+                ].spacing(10);
+
+
+        let cancel_button = button(text("Abbrechen").size(frontend::HEADING_TEXT_SIZE))
+            .on_press(Message::Cancel)
+            .padding(15);
+
+        let edit_button = button(row![frontend::font::edit_icon(), text("Bearbeiten").size(frontend::HEADING_TEXT_SIZE)].align_y(Center))
+            .on_press_with(move || {
+                Message::Edit(lokbuch.state.selected_lok_id.clone().unwrap())
+            })
+            .padding(15);
+
+        let remove_button = button(row![frontend::font::delete_icon(), text("Löschen").size(frontend::HEADING_TEXT_SIZE)].align_y(Center))
+            .on_press_with(move || {
+                Message::Remove(lokbuch.state.selected_lok_id.clone().unwrap())
+            })
+            .style(button::danger)
+            .padding(15);
+
+        iced::widget::column![
+                    header,
+                    row![
+                        horizontal_space(),
+                        left_column,
+                        horizontal_space(),
+                        right_column,
+                        horizontal_space(),
+                    ].width(Fill).height(Fill),
+                    center(row![
+                        horizontal_space(),
+                        edit_button,
+                        horizontal_space(),
+                        remove_button,
+                        horizontal_space(),
+                        cancel_button,
+                        horizontal_space(),
+                    ])].width(Fill).into()
+    }
+}
